+++ src/GetLanguages.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Horde\Core\Translation\Middleware\Api;
+namespace Horde\Core\Translation;
 
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ResponseInterface;
